// src/user/types/aula.response.ts
// Generated by https://quicktype.io

import type { DocenteAulaResponse } from "./docente_aula.response";

export interface AulaResponse {
    total: number;
    pages: number;
    aulas: Aula[];
}

export interface Aula {
    id: string;
    name: string;
    description: string;
    capacity: number;
    state: state;
    docenteAula: DocenteAulaResponse[];
}

export type state = "available" | "maintenance" | "busy";